vantage6-backend-common/requirements.txt
flask==2.2.5
flask-mail==0.9.1
Flask-RESTful==0.3.10
marshmallow==3.19.0
marshmallow-sqlalchemy==1.4.0
sqlalchemy==2.0.37
python-keycloak==5.5.0